WIP: Added intensity enumerator and implemented Resting class
the intensity enumerator is for the training state where you can specify how hard you want to train. The harder you train, the higher the stat reward, but the longer it takes to train. I've written out comments as to how to handle each method within the Resting method. These are general purpose comments and are not specific to the Resting class. This is for reference later.